relocation signout functionality to profile page
