+++ src/styles.scss
@@ -29,7 +29,8 @@ img {
 .comic-grid-container,
 .showcase-container,
 .fav-character-container,
-.slider-container {
+.slider-container,
+.profile-container {
 	transform-style: preserve-3d;
 	.card {
 		color: #fff;
